Fix test and indent
